+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
-## 0.1.0 (upcoming)
+## 0.17.0-0.3.0 (Upcoming)
+
+* Add clusterAPI capabilities for AWS VMs
+* Add EKS secrets encryption support
+* Fix wait conditions for unmanaged clusters
+
+## 0.17.0-0.2.0 (2023-07-03)
+
+* Add clusterAPI capabilities for AKS
+* Add clusterAPI capabilities for Azure VMs
+
+## 0.17.0-0.1.0 (2023-03-31)
 
 * Add clusterAPI capabilities for EKS
+
+## Previous development